Improved BVH tree construction performance.
git-svn-id: https://svn.syssoft.uni-trier.de/svn/prisma/DeltaWorks/include@491 87af6b1e-9b34-e211-997a-bc305bce1b9a

+++ math/graph.cpp
@@ -95,7 +95,7 @@ namespace Math
 		Queue<PNode>	queue;
 		foreach (nodes,n)
 		{
-			(*n)->inDegree = (*n)->in.count();
+			(*n)->inDegree = (int)(*n)->in.count();
 			if (!(*n)->inDegree)
 				queue.push(*n);
 		}
